+++ src/main/java/com/leets/xcellentbe/domain/follow/domain/Follow.java
@@ -1,6 +1,6 @@
 package com.leets.xcellentbe.domain.follow.domain;
 
-import com.leets.xcellentbe.domain.user.User;
+import com.leets.xcellentbe.domain.user.domain.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
