+++ src/main/java/com/leets/xcellentbe/domain/user/domain/repository/UserRepository.java
@@ -1,10 +1,10 @@
-package com.leets.xcellentbe.domain.user.repository;
+package com.leets.xcellentbe.domain.user.domain.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.leets.xcellentbe.domain.user.User;
+import com.leets.xcellentbe.domain.user.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(String email);
