+++ src/main/java/com/leets/xcellentbe/global/auth/login/LoginService.java
@@ -2,9 +2,9 @@ package com.leets.xcellentbe.global.auth.login;
 
 import static org.springframework.security.core.userdetails.User.*;
 
-import com.leets.xcellentbe.domain.user.User;
+import com.leets.xcellentbe.domain.user.domain.User;
 import com.leets.xcellentbe.domain.user.exception.UserNotFoundException;
-import com.leets.xcellentbe.domain.user.repository.UserRepository;
+import com.leets.xcellentbe.domain.user.domain.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
